+++ packages/core/src/index.ts
@@ -14,5 +14,19 @@ export {
   isValidHttpUrl,
 } from "./validators.js";
 export { bootstrapViewerToUser, isAppBootstrap, isAppLocale } from "./bootstrap.js";
+export {
+  isValidIanaTimezone,
+  localDateTimeWithTimezoneToUtcIso,
+  datePartFromUtcInstantInTimezone,
+  deriveUtcFromTemporalInput,
+  deriveAllDayEndAtUtc,
+  deriveEventEndAtUtc,
+  deriveEventUtcRange,
+} from "./temporal.js";
 export type { AppBootstrap, AppLocale, BootstrapUser, BootstrapViewer } from "./bootstrap.js";
+export type {
+  DeriveUtcFromTemporalInputOptions,
+  DeriveEventEndAtUtcOptions,
+  DerivedEventUtcRange,
+} from "./temporal.js";
 export type { SsrEventData, SsrInitialData, SsrProfileData } from "./ssr.js";
